refactor: change worldStartup to onWorldStartup
for consistency with onReload and onStart
