+++ Canopy[BP]/scripts/src/onReload.js
@@ -2,7 +2,7 @@ import { world } from '@minecraft/server';
 import { broadcastActionBar } from "../include/utils";
 import { LegacyInventoryMigrator } from "./classes/simplayer/LegacyInventoryMigrator";
 import Understudies from "./classes/simplayer/Understudies";
-import { startWorldSystems } from "./worldStartup";
+import { startWorldSystems } from "./onWorldStartup";
 
 world.afterEvents.worldLoad.subscribe(async () => {
     await LegacyInventoryMigrator.migrate();
