+++ Canopy[BP]/scripts/src/onWorldStartup.js
